# swagger_schema.json
{"openapi":"3.1.0","info":{"title":"Guppy Funds API","description":"\n## Personal Finance Ingestion and Enrichment Service\n\nAutomated system for importing, processing, and enriching financial transactions from multiple banks.\n\n### Features\n- **Multi-Bank Support**: AMEX, Citi, Wells Fargo\n- **AI-Powered Enrichment**: Claude API for merchant normalization and categorization\n- **Duplicate Detection**: Intelligent deduplication across uploads\n- **Merchant Caching**: Reuses enrichment results for cost optimization\n- **Automated Processing**: Background worker handles everything\n\n### Quick Start\n1. Upload CSV: `POST /v1/imports`\n2. Wait 30-60 seconds for automatic processing\n3. Query enriched data: `GET /v1/transactions`\n\n### Transaction Types\n- **income**: Payroll, direct deposits, salary\n- **debit**: Spending (purchases, bills, fees)\n- **credit**: Refunds, credits from merchants\n    ","version":"1.0.0"},"paths":{"/v1/imports":{"post":{"tags":["Import Management"],"summary":"Upload CSV file","description":"Upload a financial CSV file for processing.\n\n    **Supported Banks:**\n    - `amex` - American Express\n    - `citi` - Citi Bank (Costco Visa)\n    - `wells` - Wells Fargo\n\n    **What Happens:**\n    1. File is validated and saved\n    2. Import record created with state \"uploaded\"\n    3. Background worker automatically parses and enriches\n    4. Process completes in 30-60 seconds\n\n    **Duplicate Protection:**\n    - Uploading the same file multiple times will skip duplicate transactions\n    - Safe to upload overlapping monthly statements","operationId":"upload_import_v1_imports_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_import_v1_imports_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponse"}}}},"404":{"description":"Import not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/imports/{import_id}/parse":{"post":{"tags":["Import Management"],"summary":"Manually trigger parsing","description":"Manually trigger CSV parsing (normally handled by worker).\n\n    Converts CSV rows into structured transactions in the `transactions_raw` collection.","operationId":"parse_import_v1_imports__import_id__parse_post","parameters":[{"name":"import_id","in":"path","required":true,"schema":{"type":"string","title":"Import Id"}}],"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParseResponse"}}}},"404":{"description":"Import not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/imports/{import_id}/enrich":{"post":{"tags":["Import Management"],"summary":"Manually trigger enrichment","description":"Manually trigger AI enrichment (normally handled by worker).\n\n    Uses Claude API to:\n    - Normalize merchant names\n    - Assign categories and subcategories\n    - Extract location data\n    - Detect subscriptions and income\n    - Generate relevant tags\n\n    Results are cached in the `merchants` collection for future reuse.","operationId":"enrich_import_v1_imports__import_id__enrich_post","parameters":[{"name":"import_id","in":"path","required":true,"schema":{"type":"string","title":"Import Id"}}],"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrichResponse"}}}},"404":{"description":"Import not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"Health check endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check for monitoring.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_import_v1_imports_post":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"CSV file from your bank"},"source_bank":{"type":"string","enum":["amex","citi","wells"],"title":"Source Bank","description":"Bank source: amex, citi, or wells"}},"type":"object","required":["file","source_bank"],"title":"Body_upload_import_v1_imports_post"},"EnrichResponse":{"properties":{"import_id":{"type":"string","title":"Import Id"},"state":{"type":"string","title":"State"},"enriched_rows":{"type":"integer","title":"Enriched Rows","description":"Successfully enriched rows"},"failed_rows":{"type":"integer","title":"Failed Rows","description":"Failed rows"},"duration_ms":{"type":"integer","title":"Duration Ms","description":"Enrichment duration in milliseconds"},"message":{"type":"string","title":"Message"}},"type":"object","required":["import_id","state","enriched_rows","failed_rows","duration_ms","message"],"title":"EnrichResponse","description":"Response for enrichment completion."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ParseResponse":{"properties":{"import_id":{"type":"string","title":"Import Id"},"state":{"type":"string","title":"State"},"total_rows":{"type":"integer","title":"Total Rows","description":"Total rows in CSV"},"parsed_rows":{"type":"integer","title":"Parsed Rows","description":"Successfully parsed rows"},"failed_rows":{"type":"integer","title":"Failed Rows","description":"Failed rows"},"duration_ms":{"type":"integer","title":"Duration Ms","description":"Parsing duration in milliseconds"},"message":{"type":"string","title":"Message"}},"type":"object","required":["import_id","state","total_rows","parsed_rows","failed_rows","duration_ms","message"],"title":"ParseResponse","description":"Response for parsing completion."},"UploadResponse":{"properties":{"import_id":{"type":"string","title":"Import Id","description":"Unique identifier for the import job"},"state":{"type":"string","title":"State","description":"Current state of the import"},"message":{"type":"string","title":"Message","description":"Human-readable status message"}},"type":"object","required":["import_id","state","message"],"title":"UploadResponse","description":"Response for successful file upload.","example":{"import_id":"import_2025_10_28_131943","message":"File accepted. Import queued for processing.","state":"uploaded"}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}